Address review feedback from Camille
