+++ plugins/org.eclipse.glsp.layout/pom.xml
@@ -56,7 +56,7 @@
 	<dependencies>
 		<dependency>
 			<groupId>org.eclipse.glsp</groupId>
-			<artifactId>org.eclipse.glsp.api</artifactId>
+			<artifactId>org.eclipse.glsp.server</artifactId>
 			<version>0.8.0-SNAPSHOT</version>
 		</dependency>
 		<dependency>
