+++ plugins/org.eclipse.glsp.server/src/org/eclipse/glsp/server/Registry.java
@@ -13,7 +13,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
-package org.eclipse.glsp.server.internal;
+package org.eclipse.glsp.server;
 
 import java.util.Optional;
 import java.util.Set;
