+++ plugins/org.eclipse.glsp.server/src/org/eclipse/glsp/server/actions/ActionHandler.java
@@ -21,7 +21,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 
-import org.eclipse.glsp.server.internal.Handler;
+import org.eclipse.glsp.server.Handler;
 import org.eclipse.glsp.server.model.GModelState;
 
 public interface ActionHandler extends Handler<Action> {
