+++ plugins/org.eclipse.glsp.server/src/org/eclipse/glsp/server/actions/ActionRegistry.java
@@ -15,6 +15,6 @@
  ********************************************************************************/
 package org.eclipse.glsp.server.actions;
 
-import org.eclipse.glsp.server.internal.Registry;
+import org.eclipse.glsp.server.Registry;
 
 public interface ActionRegistry extends Registry<String, Action> {}
