+++ plugins/org.eclipse.glsp.server/src/org/eclipse/glsp/server/features/directediting/ContextEditValidatorRegistry.java
@@ -15,6 +15,6 @@
  ********************************************************************************/
 package org.eclipse.glsp.server.features.directediting;
 
-import org.eclipse.glsp.server.internal.Registry;
+import org.eclipse.glsp.server.Registry;
 
 public interface ContextEditValidatorRegistry extends Registry<String, ContextEditValidator> {}
