+++ plugins/org.eclipse.glsp.server/src/org/eclipse/glsp/server/utils/LayoutUtil.java
@@ -13,7 +13,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ******************************************************************************/
-package org.eclipse.glsp.server.internal.util;
+package org.eclipse.glsp.server.utils;
 
 import static org.eclipse.glsp.server.protocol.GLSPServerException.getOrThrow;
